+++ packages/sources/react/src/components/indicators/VtmnRating/VtmnRating.tsx
@@ -1,5 +1,6 @@
-import * as React from 'react';
 import clsx from 'clsx';
+import '@vtmn/css-rating/dist/index-with-vars.css';
+import * as React from 'react';
 import { VtmnRatingSize } from './types';
 
 export interface VtmnRatingProps
@@ -84,7 +85,6 @@ export const VtmnRating = ({
   value = 0,
   showValue = false,
   onChange = undefined,
-  children,
   className,
   ...props
 }: VtmnRatingProps) => {