+++ packages/sources/react/src/components/overlays/VtmnToast/VtmnToast.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import '@vtmn/css-alert/dist/index-with-vars.css';
+import '@vtmn/css-toast/dist/index-with-vars.css';
 import clsx from 'clsx';
 import { VtmnButton } from '../../actions/VtmnButton';