+++ packages/sources/vue/src/components/index.ts
@@ -5,7 +5,12 @@ import VtmnSelect from './forms/VtmnSelect';
 import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
+import VtmnProgressbar from './indicators/VtmnProgressbar';
 import VtmnTag from './indicators/VtmnTag';
+import VtmnRating from './indicators/VtmnRating';
+import VtmnBreadcrumb from './navigation/VtmnBreadcrumb';
+import VtmnBreadcrumbItem from './navigation/VtmnBreadcrumbItem';
+import VtmnNavbar from './navigation/VtmnNavbar';
 import VtmnSearch from './navigation/VtmnSearch';
 import VtmnAlert from './overlays/VtmnAlert';
 import VtmnModal from './overlays/VtmnModal';
@@ -21,6 +26,8 @@ import VtmnToggle from './selection-controls/VtmnToggle';
 import VtmnAccordion from './structure/VtmnAccordion';
 import VtmnCard from './structure/VtmnCard';
 import VtmnDivider from './structure/VtmnDivider';
+import VtmnList from './structure/VtmnList';
+import VtmnListItem from './structure/VtmnListItem';
 import VtmnSkeleton from './structure/VtmnSkeleton';
 
 export {
@@ -31,7 +38,12 @@ export {
   VtmnBadge,
   VtmnLoader,
   VtmnPrice,
+  VtmnProgressbar,
   VtmnTag,
+  VtmnRating,
+  VtmnBreadcrumb,
+  VtmnBreadcrumbItem,
+  VtmnNavbar,
   VtmnSearch,
   VtmnAlert,
   VtmnModal,
@@ -47,5 +59,7 @@ export {
   VtmnAccordion,
   VtmnCard,
   VtmnDivider,
+  VtmnList,
+  VtmnListItem,
   VtmnSkeleton,
 };