+++ packages/sources/vue/src/components/indicators/VtmnRating/index.ts
@@ -1,0 +1,8 @@
+import { App as Application } from 'vue';
+import VtmnRating from './VtmnRating.vue';
+
+VtmnRating.install = (Vue: Application) => {
+  Vue.component(VtmnRating.name, VtmnRating);
+};
+
+export default VtmnRating;